hhhhhhhhhhhhhhhh
Former-commit-id: b59d95f2960014f99a1e62a8656e9c9e4f300983 [formerly 5ded8670c7c1df993754b6f1839ebba2dcb6f63a]
Former-commit-id: c96a741d16956fc8d10c0d6af711d0198a8ff19f

+++ Core/CommandBuffer.hpp
@@ -72,7 +72,7 @@ public:
 
 	inline ::Device* Device() const { return mDevice; }
 
-	uint32_t mTriangleCount;
+	size_t mTriangleCount;
 
 private:
 	friend class Device;
